Clear actor name on sign out
