+++ kotlin/android/app/src/main/java/dev/firezone/android/core/domain/preference/GetConfigUseCase.kt
@@ -1,7 +1,7 @@
 /* Licensed under Apache 2.0 (C) 2023 Firezone, Inc. */
 package dev.firezone.android.core.domain.preference
 
-import dev.firezone.android.core.data.PreferenceRepository
+import dev.firezone.android.core.data.Repository
 import dev.firezone.android.core.data.model.Config
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
@@ -9,7 +9,7 @@ import javax.inject.Inject
 internal class GetConfigUseCase
     @Inject
     constructor(
-        private val repository: PreferenceRepository,
+        private val repository: Repository,
     ) {
         operator fun invoke(): Flow<Config> = repository.getConfig()
