+++ kotlin/android/app/src/main/java/dev/firezone/android/features/session/ui/SessionViewModel.kt
@@ -4,7 +4,7 @@ package dev.firezone.android.features.session.ui
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.firezone.android.core.data.PreferenceRepository
+import dev.firezone.android.core.data.Repository
 import dev.firezone.android.tunnel.TunnelService.Companion.State
 import dev.firezone.android.tunnel.model.Resource
 import javax.inject.Inject
@@ -14,7 +14,7 @@ internal class SessionViewModel
     @Inject
     constructor() : ViewModel() {
         @Inject
-        internal lateinit var repo: PreferenceRepository
+        internal lateinit var repo: Repository
         private val _serviceStatusLiveData = MutableLiveData<State>()
         private val _resourcesLiveData = MutableLiveData<List<Resource>>(emptyList())
 
@@ -24,6 +24,8 @@ internal class SessionViewModel
             get() = _resourcesLiveData
 
         fun clearToken() = repo.clearToken()
+
+        fun clearActorName() = repo.clearActorName()
 
         fun getActorName() = repo.getConfigSync().actorName
     }
